data_filteration/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
easyocr>=1.7.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
tqdm>=4.65.0
datasets==4.4.1